requirements.txt
apache-airflow-providers-apache-spark
pyspark==3.5.0
dbt-postgres
great-expectations
pandas
boto3
psycopg2-binary
opencv-python-headless
pyarrow
python-dotenv